perf(frontend): reduce initial bundle slice
