+++ frontend/src/__tests__/telegramMiniAppReadOnlyManifestGuardrails.test.js
@@ -2,7 +2,7 @@ import fs from 'node:fs';
 import path from 'node:path';
 import { describe, expect, it } from 'vitest';
 
-const appSource = fs.readFileSync(path.resolve(process.cwd(), 'src/App.jsx'), 'utf8');
+const appSource = fs.readFileSync(path.resolve(process.cwd(), 'src/pages/TelegramMiniAppPatientShell.jsx'), 'utf8').replace(/\r\n/g, '\n');
 
 function sourceBetween(source, start, end) {
   const startIndex = source.indexOf(start);
@@ -36,7 +36,7 @@ describe('Telegram Mini App protected runtime guardrails', () => {
     const shellSource = sourceBetween(
       appSource,
       'function TelegramMiniAppPatientShell() {',
-      'function AppShell({ children }) {'
+      'const miniAppPageStyle = {'
     );
 
     [
